+++ python/sglang/srt/platforms/device_mixin.py
@@ -27,7 +27,10 @@ Method status annotations:
 
 import enum
 import importlib
-from typing import Any, NamedTuple, Optional
+from typing import TYPE_CHECKING, NamedTuple, Optional
+
+if TYPE_CHECKING:
+    import torch  # pyright: ignore[reportMissingImports]
 
 
 class PlatformEnum(enum.Enum):
@@ -138,7 +141,9 @@ class DeviceMixin:
         """[Active] Get total device memory in bytes."""
         raise NotImplementedError
 
-    def get_current_memory_usage(self, device: Optional[Any] = None) -> float:
+    def get_current_memory_usage(
+        self, device: Optional["torch.device"] = None
+    ) -> float:
         """[Active] Get current peak memory usage in bytes."""
         raise NotImplementedError
 
@@ -150,11 +155,11 @@ class DeviceMixin:
 
     # ---- Device management ----
 
-    def get_device(self, local_rank: int) -> Any:
+    def get_device(self, local_rank: int) -> "torch.device":
         """[Planned] Return ``torch.device`` for the given local rank."""
         raise NotImplementedError
 
-    def set_device(self, device: Any) -> None:
+    def set_device(self, device: "torch.device") -> None:
         """[Planned] Set the current device."""
         raise NotImplementedError
 
@@ -199,7 +204,7 @@ class DeviceMixin:
     @classmethod
     def inference_mode(cls):
         """[Planned] Return inference mode context manager."""
-        torch = _import_torch()
+        torch = importlib.import_module("torch")
         return torch.inference_mode(mode=True)
 
     @classmethod
@@ -209,7 +214,7 @@ class DeviceMixin:
             import random
 
             np = importlib.import_module("numpy")
-            torch = _import_torch()
+            torch = importlib.import_module("torch")
 
             random.seed(seed)
             np.random.seed(seed)
@@ -237,55 +242,3 @@ class DeviceMixin:
 
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}(device={self.device_name})"
-
-
-def _import_torch() -> Any:
-    return importlib.import_module("torch")
-
-
-class CudaDeviceMixin(DeviceMixin):
-    """CUDA implementation of the shared device operations."""
-
-    _enum: PlatformEnum = PlatformEnum.CUDA
-    device_name: str = "cuda"
-    device_type: str = "cuda"
-
-    def get_device_total_memory(self, device_id: int = 0) -> int:
-        torch = _import_torch()
-        return int(torch.cuda.get_device_properties(device_id).total_memory)
-
-    def get_current_memory_usage(self, device: Optional[Any] = None) -> float:
-        torch = _import_torch()
-        return float(torch.cuda.max_memory_allocated(device))
-
-    def get_device(self, local_rank: int) -> Any:
-        torch = _import_torch()
-        return torch.device("cuda", local_rank)
-
-    def set_device(self, device: Any) -> None:
-        torch = _import_torch()
-        torch.cuda.set_device(device)
-
-    def get_device_name(self, device_id: int = 0) -> str:
-        torch = _import_torch()
-        return str(torch.cuda.get_device_name(device_id))
-
-    def get_device_capability(self, device_id: int = 0) -> DeviceCapability:
-        torch = _import_torch()
-        major, minor = torch.cuda.get_device_capability(device_id)
-        return DeviceCapability(major, minor)
-
-    def empty_cache(self) -> None:
-        torch = _import_torch()
-        torch.cuda.empty_cache()
-
-    def synchronize(self) -> None:
-        torch = _import_torch()
-        torch.cuda.synchronize()
-
-    def get_available_memory(self, device_id: int = 0) -> tuple[int, int]:
-        torch = _import_torch()
-        return torch.cuda.mem_get_info(device_id)
-
-    def get_torch_distributed_backend_str(self) -> str:
-        return "nccl"
